+++ PGLib/src/patchers/PatcherMeshShaderTruePBR.cpp
@@ -540,10 +540,7 @@ auto PatcherMeshShaderTruePBR::applyOnePatch(NiShape* nifShape, nlohmann::json& 
 
     // "delete" attribute
     if (truePBRData.contains("delete") && truePBRData["delete"]) {
-        if (nifShader->GetAlpha() > 0.0) {
-            nifShader->SetAlpha(0.0);
-            changed = true;
-        }
+        getNIF()->DeleteShape(nifShape);
         return changed;
     }